% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_review.R
\name{aiReviewPanelUI}
\alias{aiReviewPanelUI}
\title{AI Review Panel UI}
\usage{
aiReviewPanelUI(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A Shiny UI element.
}
\description{
Shiny module UI for reviewing AI-generated chunks.
}
