% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_architect.R
\docType{data}
\name{ARCHITECT_PROMPT_TEMPLATE}
\alias{ARCHITECT_PROMPT_TEMPLATE}
\title{Architect Agent for Team Composition}
\format{
An object of class \code{character} of length 1.
}
\usage{
ARCHITECT_PROMPT_TEMPLATE
}
\description{
The Architect is a meta-agent that analyzes user tasks and selects
the optimal combination of agents to accomplish the task. It acts as
an intelligent "HR system" for automatic team assembly.
}
\keyword{datasets}
