% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_architect_v2.R
\name{ArchitectV2}
\alias{ArchitectV2}
\title{Architect V2 R6 Class}
\description{
Architect V2 R6 Class

Architect V2 R6 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{agent}}{The underlying Agent object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArchitectV2-new}{\code{ArchitectV2$new()}}
\item \href{#method-ArchitectV2-plan}{\code{ArchitectV2$plan()}}
\item \href{#method-ArchitectV2-clone}{\code{ArchitectV2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchitectV2-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArchitectV2-new}{}}}
\subsection{Method \code{new()}}{
Initialize Architect V2
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchitectV2$new(capabilities_text, model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{capabilities_text}}{Formatted capabilities table}

\item{\code{model}}{Model to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchitectV2-plan"></a>}}
\if{latex}{\out{\hypertarget{method-ArchitectV2-plan}{}}}
\subsection{Method \code{plan()}}{
Plan a task with success criteria
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchitectV2$plan(task, execution_history = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{Task description}

\item{\code{execution_history}}{List of previous execution attempts}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with plan including success criteria
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchitectV2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArchitectV2-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchitectV2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
