% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_v2_helpers.R
\name{compare_genesis_results}
\alias{compare_genesis_results}
\title{Genesis V2 Helper Functions}
\usage{
compare_genesis_results(result1, result2)
}
\arguments{
\item{result1}{First result from genesis_v2()}

\item{result2}{Second result from genesis_v2()}
}
\description{
Utility functions to support Genesis V2 PER cycle
}
\keyword{Compare}
\keyword{Genesis}
\keyword{V2}
\keyword{internal}
\keyword{results}
\keyword{two}
