% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{create_team}
\alias{create_team}
\title{Create an Agent Team}
\usage{
create_team(name = "AgentTeam")
}
\arguments{
\item{name}{Team name.}
}
\value{
An AgentTeam object.
}
\description{
Helper to create an AgentTeam.
}
