% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{debug_log}
\alias{debug_log}
\title{Debug Log Helper}
\usage{
debug_log(context, data)
}
\arguments{
\item{context}{A string describing the context.}

\item{data}{A list of data to log.}
}
\description{
Log debug information if debug mode is enabled.
}
\keyword{internal}
