% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_deepseek.R
\name{DeepSeekLanguageModel}
\alias{DeepSeekLanguageModel}
\title{DeepSeek Language Model Class}
\description{
Language model implementation for DeepSeek's chat completions API.
Inherits from OpenAI model but adds support for DeepSeek-specific features
like reasoning content extraction from \code{deepseek-reasoner} model.
}
\keyword{internal}
\section{Super classes}{
\code{\link[aisdk:LanguageModelV1]{aisdk::LanguageModelV1}} -> \code{\link[aisdk:OpenAILanguageModel]{aisdk::OpenAILanguageModel}} -> \code{DeepSeekLanguageModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DeepSeekLanguageModel-do_generate}{\code{DeepSeekLanguageModel$do_generate()}}
\item \href{#method-DeepSeekLanguageModel-clone}{\code{DeepSeekLanguageModel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="LanguageModelV1" data-id="generate"><a href='../../aisdk/html/LanguageModelV1.html#method-LanguageModelV1-generate'><code>aisdk::LanguageModelV1$generate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="LanguageModelV1" data-id="stream"><a href='../../aisdk/html/LanguageModelV1.html#method-LanguageModelV1-stream'><code>aisdk::LanguageModelV1$stream()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="do_stream"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-do_stream'><code>aisdk::OpenAILanguageModel$do_stream()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="format_tool_result"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-format_tool_result'><code>aisdk::OpenAILanguageModel$format_tool_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="get_config"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-get_config'><code>aisdk::OpenAILanguageModel$get_config()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="get_history_format"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-get_history_format'><code>aisdk::OpenAILanguageModel$get_history_format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="OpenAILanguageModel" data-id="initialize"><a href='../../aisdk/html/OpenAILanguageModel.html#method-OpenAILanguageModel-initialize'><code>aisdk::OpenAILanguageModel$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeepSeekLanguageModel-do_generate"></a>}}
\if{latex}{\out{\hypertarget{method-DeepSeekLanguageModel-do_generate}{}}}
\subsection{Method \code{do_generate()}}{
Generate text (non-streaming).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeepSeekLanguageModel$do_generate(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of call options including messages, temperature, etc.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeepSeekLanguageModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DeepSeekLanguageModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeepSeekLanguageModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
