% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{deprecation_warning}
\alias{deprecation_warning}
\title{Deprecation Warning Helper}
\usage{
deprecation_warning(old_fn, new_fn, version = "2.0.0")
}
\arguments{
\item{old_fn}{Name of the deprecated function/pattern.}

\item{new_fn}{Name of the replacement.}

\item{version}{Version when it will be removed.}
}
\description{
Issue a deprecation warning with migration guidance.
}
\keyword{internal}
