% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_model.R
\name{EmbeddingModelV1}
\alias{EmbeddingModelV1}
\title{Embedding Model V1 (Abstract Base Class)}
\description{
Abstract interface for embedding models.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{specification_version}}{The version of this specification.}

\item{\code{provider}}{The provider identifier.}

\item{\code{model_id}}{The model identifier.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EmbeddingModelV1-new}{\code{EmbeddingModelV1$new()}}
\item \href{#method-EmbeddingModelV1-do_embed}{\code{EmbeddingModelV1$do_embed()}}
\item \href{#method-EmbeddingModelV1-clone}{\code{EmbeddingModelV1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmbeddingModelV1-new"></a>}}
\if{latex}{\out{\hypertarget{method-EmbeddingModelV1-new}{}}}
\subsection{Method \code{new()}}{
Initialize the embedding model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbeddingModelV1$new(provider, model_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{provider}}{Provider name.}

\item{\code{model_id}}{Model ID.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmbeddingModelV1-do_embed"></a>}}
\if{latex}{\out{\hypertarget{method-EmbeddingModelV1-do_embed}{}}}
\subsection{Method \code{do_embed()}}{
Generate embeddings for a value. Abstract method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbeddingModelV1$do_embed(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A character string or vector to embed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with embeddings.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmbeddingModelV1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EmbeddingModelV1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbeddingModelV1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
