% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_evaluator.R
\name{Evaluator}
\alias{Evaluator}
\title{Evaluator R6 Class}
\description{
Evaluator R6 Class

Evaluator R6 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{agent}}{The underlying Agent object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Evaluator-new}{\code{Evaluator$new()}}
\item \href{#method-Evaluator-evaluate}{\code{Evaluator$evaluate()}}
\item \href{#method-Evaluator-clone}{\code{Evaluator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-new}{}}}
\subsection{Method \code{new()}}{
Initialize Evaluator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$new(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Model to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
Evaluate a task result
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$evaluate(task, success_criteria, result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{Original task description}

\item{\code{success_criteria}}{Success criteria from Architect}

\item{\code{result}}{Execution result to evaluate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with evaluation details
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Evaluator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Evaluator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Evaluator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
