% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_evaluator.R
\docType{data}
\name{EVALUATOR_PROMPT}
\alias{EVALUATOR_PROMPT}
\title{Evaluator Agent: Result Quality Assessment}
\format{
An object of class \code{character} of length 1.
}
\usage{
EVALUATOR_PROMPT
}
\description{
The Evaluator Agent assesses the quality of task execution results
against predefined success criteria. This is the core of the Refine phase.
}
\keyword{datasets}
