% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_engine.R
\name{execute_code_safely}
\alias{execute_code_safely}
\title{Execute Code Safely}
\usage{
execute_code_safely(code, envir)
}
\arguments{
\item{code}{The R code to execute.}

\item{envir}{The environment in which to execute.}
}
\value{
A character string with the output or error message.
}
\description{
Executes R code in the given environment and captures output.
}
\keyword{internal}
