% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis.R
\name{execute_with_plan}
\alias{execute_with_plan}
\title{Execute task with a given plan}
\usage{
execute_with_plan(task, library, plan, model, verbose)
}
\description{
Execute task with a given plan
}
\keyword{internal}
