% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_v2_helpers.R
\name{export_genesis_result}
\alias{export_genesis_result}
\title{Export Genesis V2 result to JSON}
\usage{
export_genesis_result(result, file)
}
\arguments{
\item{result}{Result from genesis_v2()}

\item{file}{Output file path}
}
\description{
Export Genesis V2 result to JSON
}
