% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_v2.R
\name{find_best_result}
\alias{find_best_result}
\title{Find best result from execution history}
\usage{
find_best_result(execution_history)
}
\description{
Find best result from execution history
}
\keyword{internal}
