% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{find_tool}
\alias{find_tool}
\title{Find Tool by Name}
\usage{
find_tool(tools, name)
}
\arguments{
\item{tools}{A list of Tool objects.}

\item{name}{The tool name to find.}
}
\value{
The Tool object if found, NULL otherwise.
}
\description{
Find a tool in a list of tools by its name.
}
\keyword{internal}
