% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_gemini.R
\name{GeminiImageModel}
\alias{GeminiImageModel}
\title{Gemini Image Model Class}
\description{
Image model implementation for Gemini image generation models.
}
\keyword{internal}
\section{Super class}{
\code{\link[aisdk:ImageModelV1]{aisdk::ImageModelV1}} -> \code{GeminiImageModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeminiImageModel-new}{\code{GeminiImageModel$new()}}
\item \href{#method-GeminiImageModel-do_generate_image}{\code{GeminiImageModel$do_generate_image()}}
\item \href{#method-GeminiImageModel-do_edit_image}{\code{GeminiImageModel$do_edit_image()}}
\item \href{#method-GeminiImageModel-clone}{\code{GeminiImageModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="edit_image"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-edit_image'><code>aisdk::ImageModelV1$edit_image()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="generate_image"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-generate_image'><code>aisdk::ImageModelV1$generate_image()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="ImageModelV1" data-id="has_capability"><a href='../../aisdk/html/ImageModelV1.html#method-ImageModelV1-has_capability'><code>aisdk::ImageModelV1$has_capability()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeminiImageModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeminiImageModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Gemini image model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeminiImageModel$new(model_id, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_id}}{The model ID.}

\item{\code{config}}{Configuration list with api_key, base_url, headers, etc.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeminiImageModel-do_generate_image"></a>}}
\if{latex}{\out{\hypertarget{method-GeminiImageModel-do_generate_image}{}}}
\subsection{Method \code{do_generate_image()}}{
Generate images.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeminiImageModel$do_generate_image(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of call options.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateImageResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeminiImageModel-do_edit_image"></a>}}
\if{latex}{\out{\hypertarget{method-GeminiImageModel-do_edit_image}{}}}
\subsection{Method \code{do_edit_image()}}{
Edit images.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeminiImageModel$do_edit_image(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of call options.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateImageResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeminiImageModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeminiImageModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeminiImageModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
