% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_v2.R
\name{generate_success_criteria}
\alias{generate_success_criteria}
\title{Generate success criteria for a task}
\usage{
generate_success_criteria(task, model)
}
\description{
Generate success criteria for a task
}
\keyword{internal}
