% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_v2_helpers.R
\name{genesis_v2_stats}
\alias{genesis_v2_stats}
\title{Get Genesis V2 statistics}
\usage{
genesis_v2_stats(results)
}
\arguments{
\item{results}{List of results from multiple genesis_v2() calls}
}
\description{
Get Genesis V2 statistics
}
