% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_memory.R
\name{get_memory}
\alias{get_memory}
\title{Get or Create Global Memory}
\usage{
get_memory()
}
\value{
A ProjectMemory object.
}
\description{
Get the global project memory instance, creating it if necessary.
}
