% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_env.R
\name{get_openai_base_url}
\alias{get_openai_base_url}
\title{Get OpenAI Base URL from environment}
\usage{
get_openai_base_url()
}
\value{
Base URL string
}
\description{
Get OpenAI Base URL from environment
}
