% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_env.R
\name{get_openai_embedding_model}
\alias{get_openai_embedding_model}
\title{Get OpenAI Embedding Model from environment}
\usage{
get_openai_embedding_model()
}
\value{
Model name string
}
\description{
Get OpenAI Embedding Model from environment
}
