% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_env.R
\name{get_openai_model_id}
\alias{get_openai_model_id}
\title{Get OpenAI Model ID from environment}
\usage{
get_openai_model_id()
}
\value{
Model ID string
}
\description{
Get OpenAI Model ID from environment
}
