% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_generator.R
\name{get_param_docs}
\alias{get_param_docs}
\title{Look up a function's parameter documentation}
\usage{
get_param_docs(func_name, package = NULL)
}
\arguments{
\item{func_name}{Name of the function whose parameter docs to read.}

\item{package}{Optional package name containing the function.}
}
\value{
A named character vector of parameter descriptions.
}
\description{
Part of the companion-package extension API (used by \pkg{aisdk.datatools}).
}
\keyword{internal}
