% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_model.R
\name{ImageModelV1}
\alias{ImageModelV1}
\title{Image Model V1 (Abstract Base Class)}
\description{
Abstract interface for image generation and editing models.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{specification_version}}{The version of this specification.}

\item{\code{provider}}{The provider identifier.}

\item{\code{model_id}}{The model identifier.}

\item{\code{capabilities}}{Model capability flags.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImageModelV1-new}{\code{ImageModelV1$new()}}
\item \href{#method-ImageModelV1-has_capability}{\code{ImageModelV1$has_capability()}}
\item \href{#method-ImageModelV1-generate_image}{\code{ImageModelV1$generate_image()}}
\item \href{#method-ImageModelV1-edit_image}{\code{ImageModelV1$edit_image()}}
\item \href{#method-ImageModelV1-do_generate_image}{\code{ImageModelV1$do_generate_image()}}
\item \href{#method-ImageModelV1-do_edit_image}{\code{ImageModelV1$do_edit_image()}}
\item \href{#method-ImageModelV1-clone}{\code{ImageModelV1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageModelV1-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImageModelV1-new}{}}}
\subsection{Method \code{new()}}{
Initialize the image model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageModelV1$new(provider, model_id, capabilities = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{provider}}{Provider name.}

\item{\code{model_id}}{Model ID.}

\item{\code{capabilities}}{Optional list of capability flags.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageModelV1-has_capability"></a>}}
\if{latex}{\out{\hypertarget{method-ImageModelV1-has_capability}{}}}
\subsection{Method \code{has_capability()}}{
Check if the image model has a specific capability.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageModelV1$has_capability(cap)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cap}}{Capability name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageModelV1-generate_image"></a>}}
\if{latex}{\out{\hypertarget{method-ImageModelV1-generate_image}{}}}
\subsection{Method \code{generate_image()}}{
Public image generation method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageModelV1$generate_image(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Call options passed to \code{do_generate_image()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateImageResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageModelV1-edit_image"></a>}}
\if{latex}{\out{\hypertarget{method-ImageModelV1-edit_image}{}}}
\subsection{Method \code{edit_image()}}{
Public image editing method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageModelV1$edit_image(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Call options passed to \code{do_edit_image()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateImageResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageModelV1-do_generate_image"></a>}}
\if{latex}{\out{\hypertarget{method-ImageModelV1-do_generate_image}{}}}
\subsection{Method \code{do_generate_image()}}{
Generate images. Abstract method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageModelV1$do_generate_image(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of call options.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateImageResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageModelV1-do_edit_image"></a>}}
\if{latex}{\out{\hypertarget{method-ImageModelV1-do_edit_image}{}}}
\subsection{Method \code{do_edit_image()}}{
Edit images. Abstract method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageModelV1$do_edit_image(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of call options.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A GenerateImageResult object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImageModelV1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImageModelV1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImageModelV1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
