% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_v2_helpers.R
\name{import_genesis_result}
\alias{import_genesis_result}
\title{Import Genesis V2 result from JSON}
\usage{
import_genesis_result(file)
}
\arguments{
\item{file}{Input file path}
}
\description{
Import Genesis V2 result from JSON
}
