% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcp.R
\name{jsonrpc_response}
\alias{jsonrpc_response}
\title{Create a JSON-RPC 2.0 success response object}
\usage{
jsonrpc_response(result, id)
}
\arguments{
\item{result}{The result value}

\item{id}{The request ID this is responding to}
}
\value{
A list representing the JSON-RPC response
}
\description{
Create a JSON-RPC 2.0 success response object
}
\keyword{internal}
