% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mcp.R
\name{mcp_tools_list_request}
\alias{mcp_tools_list_request}
\title{Create MCP tools/list request}
\usage{
mcp_tools_list_request(id)
}
\arguments{
\item{id}{Request ID}
}
\value{
A JSON-RPC request
}
\description{
Create MCP tools/list request
}
\keyword{internal}
