% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_sse.R
\name{McpSseClient}
\alias{McpSseClient}
\title{MCP SSE Client}
\description{
Connect to an MCP server via Server-Sent Events (SSE).
}
\details{
Manages connection to a remote MCP server via SSE transport.
}
\section{Super class}{
\code{\link[aisdk:McpClient]{aisdk::McpClient}} -> \code{McpSseClient}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{endpoint}}{The POST endpoint for sending messages (received from SSE init)}

\item{\code{auth_headers}}{Authentication headers}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-McpSseClient-new}{\code{McpSseClient$new()}}
\item \href{#method-McpSseClient-clone}{\code{McpSseClient$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="McpClient" data-id="as_sdk_tools"><a href='../../aisdk/html/McpClient.html#method-McpClient-as_sdk_tools'><code>aisdk::McpClient$as_sdk_tools()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="McpClient" data-id="call_tool"><a href='../../aisdk/html/McpClient.html#method-McpClient-call_tool'><code>aisdk::McpClient$call_tool()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="McpClient" data-id="close"><a href='../../aisdk/html/McpClient.html#method-McpClient-close'><code>aisdk::McpClient$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="McpClient" data-id="is_alive"><a href='../../aisdk/html/McpClient.html#method-McpClient-is_alive'><code>aisdk::McpClient$is_alive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="McpClient" data-id="list_resources"><a href='../../aisdk/html/McpClient.html#method-McpClient-list_resources'><code>aisdk::McpClient$list_resources()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="McpClient" data-id="list_tools"><a href='../../aisdk/html/McpClient.html#method-McpClient-list_tools'><code>aisdk::McpClient$list_tools()</code></a></span></li>
<li><span class="pkg-link" data-pkg="aisdk" data-topic="McpClient" data-id="read_resource"><a href='../../aisdk/html/McpClient.html#method-McpClient-read_resource'><code>aisdk::McpClient$read_resource()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-McpSseClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-McpSseClient-new}{}}}
\subsection{Method \code{new()}}{
Create a new MCP SSE Client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{McpSseClient$new(url, headers = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The SSE endpoint URL}

\item{\code{headers}}{named list of headers (e.g. for auth)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new McpSseClient object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-McpSseClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-McpSseClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{McpSseClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
