% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_architect.R
\name{parse_architect_response}
\alias{parse_architect_response}
\title{Parse Architect's JSON response}
\usage{
parse_architect_response(response)
}
\arguments{
\item{response}{Character string from Architect}
}
\value{
List with task_analysis, selected_agents, reasoning, delegation_strategy
}
\description{
Parse Architect's JSON response
}
