% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_architect_v2.R
\name{parse_architect_v2_response}
\alias{parse_architect_v2_response}
\title{Parse Architect V2's JSON response}
\usage{
parse_architect_v2_response(response)
}
\arguments{
\item{response}{Character string from Architect V2}
}
\value{
List with task_analysis, selected_agents, reasoning, delegation_strategy, success_criteria
}
\description{
Parse Architect V2's JSON response
}
