% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_evaluator.R
\name{parse_evaluator_response}
\alias{parse_evaluator_response}
\title{Parse Evaluator's JSON response}
\usage{
parse_evaluator_response(response)
}
\arguments{
\item{response}{Character string from Evaluator}
}
\value{
List with evaluation details
}
\description{
Parse Evaluator's JSON response
}
