% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_refiner.R
\name{parse_refiner_response}
\alias{parse_refiner_response}
\title{Parse Refiner's JSON response}
\usage{
parse_refiner_response(response)
}
\arguments{
\item{response}{Character string from Refiner}
}
\value{
List with refinement strategy
}
\description{
Parse Refiner's JSON response
}
