% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_v2_helpers.R
\name{plot_genesis_progression}
\alias{plot_genesis_progression}
\title{Visualize Genesis V2 score progression}
\usage{
plot_genesis_progression(result)
}
\arguments{
\item{result}{Result from genesis_v2()}
}
\description{
Visualize Genesis V2 score progression
}
