% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{print.aisdk_cost_estimate}
\alias{print.aisdk_cost_estimate}
\title{Print an aisdk cost estimate}
\usage{
\method{print}{aisdk_cost_estimate}(x, ...)
}
\arguments{
\item{x}{An \code{aisdk_cost_estimate}.}

\item{...}{Ignored.}
}
\description{
Print an aisdk cost estimate
}
