% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_evals.R
\name{print.aisdk_eval_suite_result}
\alias{print.aisdk_eval_suite_result}
\title{Print an Eval Suite Result}
\usage{
\method{print}{aisdk_eval_suite_result}(x, ...)
}
\arguments{
\item{x}{An \code{aisdk_eval_suite_result}.}

\item{...}{Ignored.}
}
\description{
Print an Eval Suite Result
}
