% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_evals.R
\name{print.benchmark_result}
\alias{print.benchmark_result}
\title{Print Benchmark Result}
\usage{
\method{print}{benchmark_result}(x, ...)
}
\arguments{
\item{x}{A benchmark result object.}

\item{...}{Additional arguments (not used).}
}
\description{
Print Benchmark Result
}
