% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_object.R
\name{print.GenerateObjectResult}
\alias{print.GenerateObjectResult}
\title{Print GenerateObjectResult}
\usage{
\method{print}{GenerateObjectResult}(x, ...)
}
\arguments{
\item{x}{A GenerateObjectResult object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print GenerateObjectResult
}
