% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{print.z_schema}
\alias{print.z_schema}
\title{Print Method for z_schema}
\usage{
\method{print}{z_schema}(x, ...)
}
\arguments{
\item{x}{A z_schema object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Pretty print a z_schema object.
}
