% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_v2.R
\name{print_genesis_v2_result}
\alias{print_genesis_v2_result}
\title{Print Genesis V2 result summary}
\usage{
print_genesis_v2_result(result)
}
\arguments{
\item{result}{Result object from genesis_v2()}
}
\description{
Print Genesis V2 result summary
}
