% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_evals.R
\name{private_calculate_tool_accuracy}
\alias{private_calculate_tool_accuracy}
\title{Calculate Tool Accuracy}
\usage{
private_calculate_tool_accuracy(results, tasks)
}
\description{
Calculate Tool Accuracy
}
\keyword{internal}
