% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis_refiner.R
\name{Refiner}
\alias{Refiner}
\title{Refiner R6 Class}
\description{
Refiner R6 Class

Refiner R6 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{agent}}{The underlying Agent object}

\item{\code{library}}{AgentLibrary reference}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Refiner-new}{\code{Refiner$new()}}
\item \href{#method-Refiner-refine}{\code{Refiner$refine()}}
\item \href{#method-Refiner-clone}{\code{Refiner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Refiner-new"></a>}}
\if{latex}{\out{\hypertarget{method-Refiner-new}{}}}
\subsection{Method \code{new()}}{
Initialize Refiner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Refiner$new(capabilities_text, library, model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{capabilities_text}}{Formatted capabilities table}

\item{\code{library}}{AgentLibrary instance}

\item{\code{model}}{Model to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Refiner-refine"></a>}}
\if{latex}{\out{\hypertarget{method-Refiner-refine}{}}}
\subsection{Method \code{refine()}}{
Analyze failure and generate improvement strategy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Refiner$refine(task, plan, result, evaluation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{Original task description}

\item{\code{plan}}{Execution plan that was used}

\item{\code{result}}{Execution result}

\item{\code{evaluation}}{Evaluation from Evaluator}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with refinement strategy
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Refiner-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Refiner-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Refiner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
