% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_registry.R
\name{sdk_is_var_locked}
\alias{sdk_is_var_locked}
\title{Check if a Variable is Locked}
\usage{
sdk_is_var_locked(name)
}
\arguments{
\item{name}{Character string. The name of the variable.}
}
\value{
TRUE if the variable is protected and locked, FALSE otherwise.
}
\description{
Check if a Variable is Locked
}
