% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_registry.R
\name{sdk_unprotect_var}
\alias{sdk_unprotect_var}
\title{Unprotect a Variable}
\usage{
sdk_unprotect_var(name)
}
\arguments{
\item{name}{Character string. The name of the variable to unprotect.}
}
\value{
Invisible TRUE.
}
\description{
Removes protection from a previously protected variable.
}
