% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesis.R
\name{set_skill_library}
\alias{set_skill_library}
\title{Set default skill library path}
\usage{
set_skill_library(path)
}
\arguments{
\item{path}{Character string path to skill library directory}
}
\description{
Set default skill library path
}
\examples{
\dontrun{
set_skill_library("~/my_skills")
}
}
