% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cli.R
\name{should_show_thinking}
\alias{should_show_thinking}
\title{Check if thinking content should be shown}
\usage{
should_show_thinking()
}
\description{
Check if thinking content should be shown
}
\keyword{internal}
