% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{to_snake_case}
\alias{to_snake_case}
\title{Convert to Snake Case}
\usage{
to_snake_case(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
Snake case version of the string.
}
\description{
Convert camelCase or PascalCase to snake_case.
}
\keyword{internal}
