% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ast.R
\name{walk_ast}
\alias{walk_ast}
\title{Walk an Abstract Syntax Tree}
\usage{
walk_ast(expr, visitor)
}
\arguments{
\item{expr}{An R expression, call, or primitive type.}

\item{visitor}{A function taking a node as argument.}
}
\description{
Recursively traverse an R expression and apply a visitor function to each node.
}
