% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{z_array}
\alias{z_array}
\title{Create Array Schema}
\usage{
z_array(
  items,
  description = NULL,
  nullable = FALSE,
  default = NULL,
  min_items = NULL,
  max_items = NULL
)
}
\arguments{
\item{items}{Schema for array items (created by z_* functions).}

\item{description}{Optional description of the field.}

\item{nullable}{If TRUE, allows null values.}

\item{default}{Optional default value.}

\item{min_items}{Optional minimum number of items.}

\item{max_items}{Optional maximum number of items.}
}
\value{
A list representing JSON Schema for array.
}
\description{
Create a JSON Schema for array type.
}
\examples{
z_array(z_string(), description = "List of names")
}
