% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{z_describe}
\alias{z_describe}
\title{Describe Schema}
\usage{
z_describe(schema, description)
}
\arguments{
\item{schema}{A z_schema object.}

\item{description}{The description string.}
}
\value{
The modified z_schema object.
}
\description{
Add a description to a z_schema object (pipe-friendly).
}
