% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{z_empty_object}
\alias{z_empty_object}
\title{Create Empty Object Schema}
\usage{
z_empty_object(description = NULL)
}
\arguments{
\item{description}{Optional description.}
}
\value{
A z_schema object.
}
\description{
Create a JSON Schema for an empty object \code{{}}.
}
