% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computer.R
\name{create_computer_tools}
\alias{create_computer_tools}
\title{Create Computer Tools}
\usage{
create_computer_tools(
  computer = NULL,
  working_dir = tempdir(),
  sandbox_mode = "permissive"
)
}
\arguments{
\item{computer}{Computer instance (default: create new)}

\item{working_dir}{Working directory. Defaults to \code{tempdir()}.}

\item{sandbox_mode}{Sandbox mode: "strict", "permissive", or "none"}
}
\value{
List of Tool objects
}
\description{
Create atomic tools for computer abstraction layer. These tools provide
a small set of primitives that agents can use to perform complex actions.
}
